[TEST] Added tests and improved documentation
git-svn-id: https://open-ms.svn.sourceforge.net/svnroot/open-ms/branches/msnovogen@11679 6adb6e08-d915-0410-941f-83917bcadc18

+++ source/ANALYSIS/DENOVO/MSNOVOGEN/DefaultSeeder.C
@@ -40,7 +40,7 @@ namespace OpenMS
 	DefaultSeeder::DefaultSeeder(double pm, double pmt, std::vector<const Residue*> al) :
 	Seeder(pm,pmt,al), ds_(pm, pmt, al)
 	{
-		ds_.seed(time(0));
+		ds_.seed((unsigned int)time(0));
 	}
 
 	boost::shared_ptr<Chromosome> DefaultSeeder::createIndividual() const
